fuchs/requirements.txt
"fastapi[standard]"
markdown
requests
psycopg2-binary
